fix(types): removed nestjs-i18n fake types module
