+++ global/@types/index.d.ts
@@ -5,14 +5,24 @@ export type imageURL = `blob:${URL}${string}`;
 export type URL = `${string}://${string}.${string}`;
 
 export type KeysOf<T extends object> = Array<keyof T>;
-export type KeysRecursiveOf<T> = T extends object
-	? {
-			[K in keyof T]: K extends symbol
-				? never
-				:
-						| K
-						| (T[K] extends object
-								? `${K extends symbol ? never : K}.${KeysRecursiveOf<T[K]>}`
-								: `${K extends symbol ? never : K}`);
-	  }[keyof T]
-	: '';
+
+export type IsAny<T> = unknown extends T ? ([keyof T] extends [never] ? false : true) : false;
+export type PathImpl<T, Key extends keyof T> = Key extends string
+	? IsAny<T[Key]> extends true
+		? never
+		: T[Key] extends Record<string, unknown>
+		?
+				| `${Key}.${PathImpl<T[Key], Exclude<keyof T[Key], keyof unknown[]>> & string}`
+				| `${Key}.${Exclude<keyof T[Key], keyof unknown[]> & string}`
+		: never
+	: never;
+
+export type PathImpl2<T> = PathImpl<T, keyof T> | keyof T;
+
+export type Path<T> = keyof T extends string
+	? PathImpl2<T> extends infer P
+		? P extends string | keyof T
+			? P
+			: keyof T
+		: keyof T
+	: never;
